Change order in main menu
The proposed order has been added as a comment:

1. Official documentation
2. Community documentation
3. About documenation
4. Teams

It is a little difficult to determine, what is community and what is
"official", but we propose to assume snippets, cheat sheets and external
extensions as "community" documentation.

Additionally, the snippets only contain snippets, so we remove "Tipps
and Howto" in menu title because that is misleading and not helpful.