Changes to system extension page
- use package name instead of extkey (as we now have Composer as
  preferred installation method)
- rename comments to description and add description

The comments section was previously used inconsitently, sometimes
with a description, sometimes with information about the state
of the docs or when it was added.

The description might be helpful, especially for beginners. The
text was taken from the composer.json, in some cases only the
first sentence was added.

Resolves: #177