Add information to sysext page
The system extension page lists documentation for system extensions.
But documenation for the most important extensions is missing (e.g. core).

This may be confusing and irritating for beginners. We explain where to
find the information for core, backend etc. ("TYPO3 Explained"), and
also refer to the "Extbase / Fluid book" for extbase and fluid.