Move extension manuals search to separate page
This moves the extension manuals search to a separate page and
thus fixes the problem with the automatic scrolling on that page.

Additionally, information is added about what the Extension Manual
Search will find and what it will not find.

Also, this patch makes the extension manual search a little less
prominent and propagates searching on extensions.typo3.org.

Both have their pros and cons, but it is very confusing for the
user to have to know if an extension provides documentation on
docs.typo3.org and currently many extensions do not.

The extension manuals search on docs.typo3.org had some drawbacks:

- automatic scrolling of extension page
- only third party extensions are shown which are rendered on
  docs.typo3.org (see above)
- in some cases, the results may be incorrect (see issues)

Unfortunately, we currently have no way to obtain a full list of all extension docs:

- the Extension Manual Search only shows extension rendered on docs.typo3.org
  (and no system extensions)
- the search on extensions.typo3.org only shows extensions registered on
  extensions.typo3.org and no system extensions

Resolves: #193
Related: #192
Related: #155