[TASK] Bring back system extensions submenu
With the change from commit https://github.com/TYPO3-Documentation/DocsTypo3Org-Homepage/commit/799018686756e66de7b35300c841853a7d91b794
the tiles layout was introduced and the submenu removed. To
improve the usability the submenu is introduced again. Now
one can click on the "System Extension" menu and on the
desired manual directly again without reading and scrolling
through the tiles (which costs some more seconds).

In contrast to the previous implementation (which links to master/main),
the last stable version of a manual is linked.