[TASK] Add 13-dev links to "System extensions" (#231)
Additionally, fix the spelling of "Core".

EXT:t3editor is not available anymore for v13, but was merged into EXT:backend.
The documentation page for 13-dev is linked to the new "Code editor" page TYPO3 Explained,
so users can easily find the new documentation. With the release of 13 LTS the t3editor part
can be moved down to the "abandonded" section.